Split supervisor PR lifecycle tests
