+++ .codex-supervisor/issue-journal.md
@@ -1,37 +1,37 @@
-# Issue #486: Test cleanup: split supervisor execution coverage by orchestration and cleanup boundaries
+# Issue #489: Test cleanup: split supervisor PR lifecycle coverage by readiness and review-blocker boundaries
 
 ## Supervisor Snapshot
-- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/486
-- Branch: codex/issue-486
-- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-486
-- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-486/.codex-supervisor/issue-journal.md
+- Issue URL: https://github.com/TommyKammy/codex-supervisor/issues/489
+- Branch: codex/issue-489
+- Workspace: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-489
+- Journal: /home/tommy/Dev/codex-supervisor-self-worktrees/issue-489/.codex-supervisor/issue-journal.md
 - Current phase: reproducing
 - Attempt count: 1 (implementation=1, repair=0)
-- Last head SHA: 3d7ddc8b15777802ca72436dce98f9e17dc1d5fb
+- Last head SHA: 9704168f31e53da4a7139911a933b74613c5ca4c
 - Blocked reason: none
 - Last failure signature: none
 - Repeated failure signature count: 0
-- Updated at: 2026-03-17T15:33:07.995Z
+- Updated at: 2026-03-17T15:54:52.619Z
 
 ## Latest Codex Summary
-- None yet.
+- Split `src/supervisor/supervisor-pr-lifecycle.test.ts` into readiness and review-blocker suites, preserved the legacy file as a thin import facade, and re-verified the focused supervisor PR lifecycle tests plus `npm run build`.
 
 ## Active Failure Context
 - None recorded.
 
 ## Codex Working Notes
 ### Current Handoff
-- Hypothesis: #486 can stay behavior-neutral by extracting `supervisor-execution.test.ts` into two focused suites: issue-selection/orchestration coverage and workspace/lock/cleanup-heavy execution coverage, with the legacy path preserved as a facade import.
-- What changed: split the execution suite into `src/supervisor/supervisor-execution-orchestration.test.ts` and `src/supervisor/supervisor-execution-cleanup.test.ts`, and reduced `src/supervisor/supervisor-execution.test.ts` to a thin facade import file.
+- Hypothesis: #489 can stay behavior-neutral by splitting `supervisor-pr-lifecycle.test.ts` into one readiness/post-turn suite and one review-blocker/thread-reprocessing suite, with the legacy path preserved as a facade import for existing test entry points.
+- What changed: extracted readiness coverage into `src/supervisor/supervisor-pr-readiness.test.ts`, moved review-blocker and configured-bot thread reprocessing coverage into `src/supervisor/supervisor-pr-review-blockers.test.ts`, and reduced `src/supervisor/supervisor-pr-lifecycle.test.ts` to a thin facade import file.
 - Current blocker: none
-- Next exact step: Monitor draft PR #498 for review or CI feedback and handle any verification signal on `codex/issue-486`.
-- Verification gap: none for the split execution suites or `npm run build`; `npm ci` was required first because `tsc` was initially unavailable locally.
-- Files touched: `src/supervisor/supervisor-execution.test.ts`, `src/supervisor/supervisor-execution-orchestration.test.ts`, `src/supervisor/supervisor-execution-cleanup.test.ts`, `.codex-supervisor/issue-journal.md`
-- Rollback concern: reverting this cleanup would push orchestration, session-lock, orphan cleanup, and reservation recovery assertions back into one 1.6k-line suite, making failures harder to localize.
+- Next exact step: Commit the suite split on `codex/issue-489`, then push/open the draft PR if one still does not exist.
+- Verification gap: none for the split PR lifecycle suites or `npm run build`; `npm ci` was required first because `tsc` was initially unavailable locally.
+- Files touched: `src/supervisor/supervisor-pr-lifecycle.test.ts`, `src/supervisor/supervisor-pr-readiness.test.ts`, `src/supervisor/supervisor-pr-review-blockers.test.ts`, `.codex-supervisor/issue-journal.md`
+- Rollback concern: reverting this cleanup would put readiness transitions, post-turn PR refresh, configured-bot thread reprocessing, and manual/local review blockers back into a single 915-line suite, making failures harder to localize.
 - Last focused command: `npm run build`
 ### Scratchpad
-- 2026-03-18: Committed the split as `c7b7f41` (`Split supervisor execution test coverage`), pushed `codex/issue-486`, and opened draft PR #498 (`https://github.com/TommyKammy/codex-supervisor/pull/498`).
-- 2026-03-18: Split `src/supervisor/supervisor-execution.test.ts` into orchestration and cleanup suites behind a facade import; focused verification was `npx tsx --test src/supervisor/supervisor-execution-orchestration.test.ts src/supervisor/supervisor-execution-cleanup.test.ts src/supervisor/supervisor-execution.test.ts`.
+- 2026-03-18: Focused reproducer for #489 was `npx tsx --test src/supervisor/supervisor-pr-lifecycle.test.ts`, which passed before the split and confirmed the cleanup was behavior-neutral rather than a failing behavior fix.
+- 2026-03-18: Verification for #489 was `npx tsx --test src/supervisor/supervisor-pr-readiness.test.ts src/supervisor/supervisor-pr-review-blockers.test.ts`, `npx tsx --test src/supervisor/supervisor-pr-lifecycle.test.ts`, then `npm ci` and `npm run build`; `npm run build` first failed with `sh: 1: tsc: not found` before dependencies were restored.
 - 2026-03-18: `npm run build` initially failed with `sh: 1: tsc: not found`; `npm ci` restored the local toolchain and the rerun of `npm run build` passed.
 - 2026-03-17: Review repair for PR #497 updates the guardrail provenance fixtures so committed shared-memory files are added before `headSha` is captured; focused verification was `npx tsx --test src/supervisor/supervisor-diagnostics-guardrail-reporting.test.ts` and `npm run build`.
 - 2026-03-17: Added `draftSkipAt` to configured-bot summaries and hydrated PRs; focused verification was `npx tsx --test src/github/github-review-signals.test.ts src/github/github-pull-request-hydrator.test.ts`, followed by `npm ci` and `npm run build`.
