+++ src/supervisor/supervisor-pr-lifecycle.test.ts
@@ -1,915 +1,2 @@
-import test from "node:test";
-import assert from "node:assert/strict";
-import fs from "node:fs/promises";
-import path from "node:path";
-import { Supervisor } from "./supervisor";
-import { formatDetailedStatus } from "./supervisor-status-rendering";
-import {
-  GitHubIssue,
-  GitHubPullRequest,
-  PullRequestCheck,
-  ReviewThread,
-  SupervisorStateFile,
-} from "../core/types";
-import {
-  branchName,
-  createConfig,
-  createRecord,
-  createReviewThread,
-  createSupervisorFixture,
-  executionReadyBody,
-  git,
-} from "./supervisor-test-helpers";
-
-async function withStubbedDateNow<T>(nowIso: string, run: () => Promise<T>): Promise<T> {
-  const originalDateNow = Date.now;
-  Date.now = () => Date.parse(nowIso);
-  try {
-    return await run();
-  } finally {
-    Date.now = originalDateNow;
-  }
-}
-
-test("runOnce marks a clean draft PR ready and enables auto-merge after the turn", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 92;
-  const branch = branchName(fixture.config, issueNumber);
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "stabilizing",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: 113,
-        blocked_reason: null,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Extract post-turn PR transitions",
-    body: "",
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const draftPr: GitHubPullRequest = {
-    number: 113,
-    title: "Post-turn transition refactor",
-    url: "https://example.test/pr/113",
-    state: "OPEN",
-    createdAt: "2026-03-13T00:10:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: "head-113",
-    mergedAt: null,
-  };
-  const readyPr: GitHubPullRequest = {
-    ...draftPr,
-    isDraft: false,
-  };
-
-  let readyCalls = 0;
-  let autoMergeCalls = 0;
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { executeCodexTurn: typeof supervisor["executeCodexTurn"] }).executeCodexTurn = async (context) => ({
-    kind: "completed",
-    record: {
-      ...context.record,
-      last_head_sha: "head-113",
-    },
-    workspaceStatus: context.workspaceStatus,
-    pr: context.pr,
-    checks: context.checks,
-    reviewThreads: context.reviewThreads,
-  });
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async (branchName: string, prNumber: number | null) => {
-      assert.equal(branchName, branch);
-      assert.equal(prNumber, 113);
-      return draftPr;
-    },
-    getChecks: async (prNumber: number) => {
-      assert.equal(prNumber, 113);
-      return [];
-    },
-    getUnresolvedReviewThreads: async (prNumber: number) => {
-      assert.equal(prNumber, 113);
-      return [];
-    },
-    getPullRequest: async (prNumber: number) => {
-      assert.equal(prNumber, 113);
-      return readyCalls === 0 ? draftPr : readyPr;
-    },
-    markPullRequestReady: async (prNumber: number) => {
-      assert.equal(prNumber, 113);
-      readyCalls += 1;
-    },
-    enableAutoMerge: async (prNumber: number, headSha: string) => {
-      assert.equal(prNumber, 113);
-      assert.equal(headSha, "head-113");
-      autoMergeCalls += 1;
-    },
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  const message = await supervisor.runOnce({ dryRun: false });
-  assert.match(message, /state=merging/);
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.pr_number, 113);
-  assert.equal(record.state, "merging");
-  assert.equal(record.last_head_sha, "head-113");
-  assert.equal(record.blocked_reason, null);
-  assert.equal(readyCalls, 1);
-  assert.equal(autoMergeCalls, 1);
-});
-
-test("runOnce waits for Copilot propagation after marking a draft PR ready", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 101;
-  const branch = branchName(fixture.config, issueNumber);
-  const config = createConfig({
-    ...fixture.config,
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "stabilizing",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: 114,
-        blocked_reason: null,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Honor the refreshed review-wait snapshot after ready for review",
-    body: "",
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const draftPr: GitHubPullRequest = {
-    number: 114,
-    title: "Propagate Copilot wait state",
-    url: "https://example.test/pr/114",
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: "head-114",
-    mergedAt: null,
-    copilotReviewState: "not_requested",
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-  const postReadyPr: GitHubPullRequest = {
-    ...draftPr,
-    isDraft: false,
-  };
-  const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-  let readyCalls = 0;
-  let autoMergeCalls = 0;
-  const supervisor = new Supervisor(config);
-  (supervisor as unknown as { executeCodexTurn: typeof supervisor["executeCodexTurn"] }).executeCodexTurn = async (context) => ({
-    kind: "completed",
-    record: context.record,
-    workspaceStatus: context.workspaceStatus,
-    pr: context.pr,
-    checks: context.checks,
-    reviewThreads: context.reviewThreads,
-  });
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async (branchName: string, prNumber: number | null) => {
-      assert.equal(branchName, branch);
-      assert.equal(prNumber, 114);
-      return draftPr;
-    },
-    getChecks: async (prNumber: number) => {
-      assert.equal(prNumber, 114);
-      return checks;
-    },
-    getUnresolvedReviewThreads: async (prNumber: number) => {
-      assert.equal(prNumber, 114);
-      return [];
-    },
-    getPullRequest: async (prNumber: number) => {
-      assert.equal(prNumber, 114);
-      return readyCalls === 0 ? draftPr : postReadyPr;
-    },
-    markPullRequestReady: async (prNumber: number) => {
-      assert.equal(prNumber, 114);
-      readyCalls += 1;
-    },
-    enableAutoMerge: async () => {
-      autoMergeCalls += 1;
-    },
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  await withStubbedDateNow("2026-03-13T06:26:22Z", async () => {
-    const message = await supervisor.runOnce({ dryRun: false });
-    assert.match(message, /state=waiting_ci/);
-  });
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.pr_number, 114);
-  assert.equal(record.state, "waiting_ci");
-  assert.equal(record.last_head_sha, "head-114");
-  assert.equal(record.review_wait_head_sha, "head-114");
-  assert.ok(record.review_wait_started_at);
-  assert.equal(Number.isNaN(Date.parse(record.review_wait_started_at ?? "")), false);
-  assert.equal(record.blocked_reason, null);
-  assert.equal(readyCalls, 1);
-  assert.equal(autoMergeCalls, 0);
-});
-
-test("handlePostTurnPullRequestTransitions refreshes PR state after marking ready", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 102;
-  const branch = branchName(fixture.config, issueNumber);
-  const state: SupervisorStateFile = {
-    activeIssueNumber: issueNumber,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "stabilizing",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: 116,
-        blocked_reason: null,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Refresh post-ready PR state",
-    body: "",
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const draftPr: GitHubPullRequest = {
-    number: 116,
-    title: "Refresh after ready",
-    url: "https://example.test/pr/116",
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: true,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: "head-116",
-    mergedAt: null,
-    copilotReviewState: "not_requested",
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-  const readyPr: GitHubPullRequest = {
-    ...draftPr,
-    isDraft: false,
-  };
-  const initialChecks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-  const postReadyChecks: PullRequestCheck[] = [{ name: "build", state: "IN_PROGRESS", bucket: "pending", workflow: "CI" }];
-
-  let readyCalls = 0;
-  let snapshotLoads = 0;
-  let syncJournalCalls = 0;
-  const supervisor = new Supervisor(fixture.config);
-  (supervisor as unknown as { loadOpenPullRequestSnapshot: (prNumber: number) => Promise<unknown> }).loadOpenPullRequestSnapshot = async (
-    prNumber: number,
-  ) => {
-    assert.equal(prNumber, 116);
-    snapshotLoads += 1;
-    return snapshotLoads === 1
-      ? { pr: draftPr, checks: initialChecks, reviewThreads: [] }
-      : { pr: readyPr, checks: postReadyChecks, reviewThreads: [] };
-  };
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    markPullRequestReady: async (prNumber: number) => {
-      assert.equal(prNumber, 116);
-      readyCalls += 1;
-    },
-  };
-
-  const result = await (
-    supervisor as unknown as {
-      handlePostTurnPullRequestTransitions: (context: {
-        state: SupervisorStateFile;
-        record: ReturnType<typeof createRecord>;
-        issue: GitHubIssue;
-        workspacePath: string;
-        syncJournal: (record: ReturnType<typeof createRecord>) => Promise<void>;
-        memoryArtifacts: { alwaysReadFiles: string[]; onDemandFiles: string[] };
-        pr: GitHubPullRequest;
-        options: { dryRun: boolean };
-      }) => Promise<{
-        record: ReturnType<typeof createRecord>;
-        pr: GitHubPullRequest;
-        checks: PullRequestCheck[];
-        reviewThreads: ReviewThread[];
-      }>;
-    }
-  ).handlePostTurnPullRequestTransitions({
-    state,
-    record: state.issues[String(issueNumber)]!,
-    issue,
-    workspacePath: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-    syncJournal: async () => {
-      syncJournalCalls += 1;
-    },
-    memoryArtifacts: { alwaysReadFiles: [], onDemandFiles: [] },
-    pr: draftPr,
-    options: { dryRun: false },
-  });
-
-  assert.equal(result.pr.isDraft, false);
-  assert.equal(result.record.state, "waiting_ci");
-  assert.equal(result.record.review_wait_head_sha, "head-116");
-  assert.ok(result.record.review_wait_started_at);
-  assert.equal(readyCalls, 1);
-  assert.equal(snapshotLoads, 2);
-  assert.equal(syncJournalCalls, 0);
-});
-
-test("runOnce records an observed Copilot request time when GitHub omits the request timestamp", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 115;
-  const branch = branchName(fixture.config, issueNumber);
-  const config = createConfig({
-    ...fixture.config,
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {},
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Persist observed Copilot request time",
-    body: executionReadyBody("Persist observed Copilot request time."),
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const pr: GitHubPullRequest = {
-    number: 115,
-    title: "Missing Copilot request timestamp",
-    url: "https://example.test/pr/115",
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: "head-115",
-    mergedAt: null,
-    copilotReviewState: "requested",
-    copilotReviewRequestedAt: null,
-    copilotReviewArrivedAt: null,
-  };
-  const checks: PullRequestCheck[] = [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }];
-
-  const supervisor = new Supervisor(config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async (branchName: string, prNumber: number | null) => {
-      assert.equal(branchName, branch);
-      assert.equal(prNumber, null);
-      return pr;
-    },
-    getChecks: async (prNumber: number) => {
-      assert.equal(prNumber, pr.number);
-      return checks;
-    },
-    getUnresolvedReviewThreads: async (prNumber: number) => {
-      assert.equal(prNumber, pr.number);
-      return [];
-    },
-    getPullRequest: async (prNumber: number) => {
-      assert.equal(prNumber, pr.number);
-      return pr;
-    },
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  await withStubbedDateNow("2026-03-13T06:26:22Z", async () => {
-    const message = await supervisor.runOnce({ dryRun: true });
-    assert.match(message, /state=waiting_ci/);
-  });
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.state, "waiting_ci");
-  assert.equal(record.pr_number, 115);
-  assert.equal(record.copilot_review_requested_head_sha, "head-115");
-  assert.ok(record.copilot_review_requested_observed_at);
-  assert.equal(Number.isNaN(Date.parse(record.copilot_review_requested_observed_at ?? "")), false);
-});
-
-test("runOnce reprocesses a configured bot review thread once after a new PR head commit and then blocks if it remains unresolved", async () => {
-  const fixture = await createSupervisorFixture({
-    codexScriptLines: [
-      "#!/bin/sh",
-      "set -eu",
-      'out=""',
-      'while [ "$#" -gt 0 ]; do',
-      '  case "$1" in',
-      '    -o) out=\"$2\"; shift 2 ;;',
-      '    *) shift ;;',
-      '  esac',
-      "done",
-      'printf \'{\"type\":\"thread.started\",\"thread_id\":\"thread-review\"}\\n\'',
-      "cat <<'EOF' > \"$out\"",
-      "Summary: reviewed configured bot thread once on the new head",
-      "State hint: stabilizing",
-      "Blocked reason: none",
-      "Tests: not run",
-      "Failure signature: none",
-      "Next action: refresh the PR snapshot and decide whether the thread still blocks",
-      "EOF",
-      "printf '\\n- Scratchpad note: review reprocessing completed for the current head.\\n' >> .codex-supervisor/issue-journal.md",
-      "exit 0",
-      "",
-    ],
-  });
-  const issueNumber = 116;
-  const branch = branchName(fixture.config, issueNumber);
-  const runHeadSha = git(["rev-parse", "HEAD"], fixture.repoPath);
-  const config = createConfig({
-    ...fixture.config,
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "pr_open",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: 116,
-        last_head_sha: "head-a",
-        processed_review_thread_ids: ["thread-1@head-a"],
-        processed_review_thread_fingerprints: ["thread-1@head-a#comment-1"],
-        blocked_reason: "manual_review",
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Reprocess configured bot review threads after a new head commit",
-    body: executionReadyBody("Reprocess configured bot review threads after a new head commit."),
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const pr: GitHubPullRequest = {
-    number: 116,
-    title: "Reprocess review threads",
-    url: "https://example.test/pr/116",
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: runHeadSha,
-    mergedAt: null,
-  };
-  const reviewThreads = [createReviewThread()];
-
-  const supervisor = new Supervisor(config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async (branchName: string, prNumber: number | null) => {
-      assert.equal(branchName, branch);
-      assert.equal(prNumber, 116);
-      return pr;
-    },
-    getChecks: async (prNumber: number) => {
-      assert.equal(prNumber, 116);
-      return [];
-    },
-    getUnresolvedReviewThreads: async (prNumber: number) => {
-      assert.equal(prNumber, 116);
-      return reviewThreads;
-    },
-    getPullRequest: async (prNumber: number) => {
-      assert.equal(prNumber, 116);
-      return pr;
-    },
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  const message = await supervisor.runOnce({ dryRun: false });
-  assert.match(message, /state=blocked/);
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.state, "blocked");
-  assert.equal(record.last_head_sha, runHeadSha);
-  assert.equal(record.blocked_reason, "manual_review");
-  assert.deepEqual(record.processed_review_thread_ids, ["thread-1@head-a", `thread-1@${runHeadSha}`]);
-  assert.deepEqual(record.processed_review_thread_fingerprints, ["thread-1@head-a#comment-1", `thread-1@${runHeadSha}#comment-1`]);
-  assert.equal(record.last_failure_context?.category, "manual");
-  assert.match(
-    record.last_failure_context?.summary ?? "",
-    /configured bot review thread\(s\) remain unresolved after processing on the current head/,
-  );
-  assert.deepEqual(record.last_failure_context?.details, [
-    "reviewer=copilot-pull-request-reviewer file=src/file.ts line=12 processed_on_current_head=yes",
-  ]);
-
-  const status = formatDetailedStatus({
-    config,
-    activeRecord: record,
-    latestRecord: record,
-    trackedIssueCount: 1,
-    pr,
-    checks: [],
-    reviewThreads,
-  });
-  assert.match(
-    status,
-    /failure_context category=manual summary=1 configured bot review thread\(s\) remain unresolved after processing on the current head and now require manual attention\./,
-  );
-  assert.match(
-    status,
-    /failure_details=reviewer=copilot-pull-request-reviewer file=src\/file\.ts line=12 processed_on_current_head=yes/,
-  );
-});
-
-test("runOnce does not mark configured bot review threads as processed for a refreshed PR head it did not evaluate", async () => {
-  const fixture = await createSupervisorFixture({
-    codexScriptLines: [
-      "#!/bin/sh",
-      "set -eu",
-      'out=""',
-      'while [ "$#" -gt 0 ]; do',
-      '  case "$1" in',
-      '    -o) out=\"$2\"; shift 2 ;;',
-      '    *) shift ;;',
-      '  esac',
-      "done",
-      'printf \'{\"type\":\"thread.started\",\"thread_id\":\"thread-review\"}\\n\'',
-      "cat <<'EOF' > \"$out\"",
-      "Summary: reviewed configured bot thread on the prior head",
-      "State hint: stabilizing",
-      "Blocked reason: none",
-      "Tests: not run",
-      "Failure signature: none",
-      "Next action: refresh the PR snapshot and decide whether a newer head still needs review",
-      "EOF",
-      "printf '\\n- Scratchpad note: review pass completed before a newer remote head appeared.\\n' >> .codex-supervisor/issue-journal.md",
-      "exit 0",
-      "",
-    ],
-  });
-  const issueNumber = 117;
-  const branch = branchName(fixture.config, issueNumber);
-  const runHeadSha = git(["rev-parse", "HEAD"], fixture.repoPath);
-  const config = createConfig({
-    ...fixture.config,
-    reviewBotLogins: ["copilot-pull-request-reviewer"],
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "pr_open",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: 117,
-        last_head_sha: "head-a",
-        processed_review_thread_ids: ["thread-1@head-a"],
-        blocked_reason: "manual_review",
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Avoid attributing review processing to an unseen head",
-    body: executionReadyBody("Avoid attributing review processing to an unseen head."),
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const initialPr: GitHubPullRequest = {
-    number: 117,
-    title: "Handle refreshed head safely",
-    url: "https://example.test/pr/117",
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: runHeadSha,
-    mergedAt: null,
-  };
-  const refreshedPr: GitHubPullRequest = {
-    ...initialPr,
-    headRefOid: "head-c",
-  };
-  const reviewThreads = [createReviewThread()];
-  let resolveCalls = 0;
-
-  const supervisor = new Supervisor(config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async (branchName: string, prNumber: number | null) => {
-      assert.equal(branchName, branch);
-      assert.equal(prNumber, 117);
-      resolveCalls += 1;
-      return resolveCalls === 1 ? initialPr : refreshedPr;
-    },
-    getChecks: async (prNumber: number) => {
-      assert.equal(prNumber, 117);
-      return [];
-    },
-    getUnresolvedReviewThreads: async (prNumber: number) => {
-      assert.equal(prNumber, 117);
-      return reviewThreads;
-    },
-    getPullRequest: async (prNumber: number) => {
-      assert.equal(prNumber, 117);
-      return refreshedPr;
-    },
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  const message = await supervisor.runOnce({ dryRun: false });
-  assert.match(message, /state=addressing_review/);
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.state, "addressing_review");
-  assert.equal(record.last_head_sha, "head-c");
-  assert.deepEqual(record.processed_review_thread_ids, ["thread-1@head-a"]);
-});
-
-test("runOnce records verification blocker context when local review blocks merge before a turn", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 118;
-  const branch = branchName(fixture.config, issueNumber);
-  const runHeadSha = git(["rev-parse", "HEAD"], fixture.repoPath);
-  const config = createConfig({
-    ...fixture.config,
-    localReviewEnabled: true,
-    localReviewPolicy: "block_merge",
-    copilotReviewWaitMinutes: 0,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "pr_open",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: issueNumber,
-        local_review_head_sha: runHeadSha,
-        local_review_findings_count: 2,
-        local_review_root_cause_count: 1,
-        local_review_max_severity: "medium",
-        local_review_recommendation: "changes_requested",
-        local_review_summary_path: "/tmp/reviews/local-review-summary.md",
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Persist local review blockers before a turn",
-    body: executionReadyBody("Persist local review blockers before a turn."),
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const pr: GitHubPullRequest = {
-    number: issueNumber,
-    title: "Preserve local review blocker context",
-    url: `https://example.test/pr/${issueNumber}`,
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: false,
-    reviewDecision: null,
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: runHeadSha,
-    mergedAt: null,
-  };
-
-  const supervisor = new Supervisor(config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async () => pr,
-    getChecks: async () => [{ name: "build", state: "SUCCESS", bucket: "pass", workflow: "CI" }],
-    getUnresolvedReviewThreads: async () => [],
-    getPullRequest: async () => pr,
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  const message = await supervisor.runOnce({ dryRun: false });
-  assert.match(message, /state=blocked/);
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.state, "blocked");
-  assert.equal(record.blocked_reason, "verification");
-  assert.match(record.last_error ?? "", /Local review found 2 actionable finding/);
-  assert.equal(record.last_failure_context?.category, "blocked");
-  assert.match(record.last_failure_context?.summary ?? "", /Local review found 2 actionable finding/);
-});
-
-test("runOnce records manual review context when GitHub reports changes requested without threads", async () => {
-  const fixture = await createSupervisorFixture();
-  const issueNumber = 119;
-  const branch = branchName(fixture.config, issueNumber);
-  const runHeadSha = git(["rev-parse", "HEAD"], fixture.repoPath);
-  const config = createConfig({
-    ...fixture.config,
-    humanReviewBlocksMerge: true,
-  });
-  const state: SupervisorStateFile = {
-    activeIssueNumber: null,
-    issues: {
-      [String(issueNumber)]: createRecord({
-        issue_number: issueNumber,
-        state: "pr_open",
-        branch,
-        workspace: path.join(fixture.workspaceRoot, `issue-${issueNumber}`),
-        journal_path: null,
-        pr_number: issueNumber,
-      }),
-    },
-  };
-  await fs.writeFile(fixture.stateFile, `${JSON.stringify(state, null, 2)}\n`, "utf8");
-
-  const issue: GitHubIssue = {
-    number: issueNumber,
-    title: "Persist changes-requested blocker context without threads",
-    body: executionReadyBody("Persist changes-requested blocker context without threads."),
-    createdAt: "2026-03-13T00:00:00Z",
-    updatedAt: "2026-03-13T00:00:00Z",
-    url: `https://example.test/issues/${issueNumber}`,
-    state: "OPEN",
-  };
-  const pr: GitHubPullRequest = {
-    number: issueNumber,
-    title: "Changes requested without threads",
-    url: `https://example.test/pr/${issueNumber}`,
-    state: "OPEN",
-    createdAt: "2026-03-13T06:20:00Z",
-    isDraft: false,
-    reviewDecision: "CHANGES_REQUESTED",
-    mergeStateStatus: "CLEAN",
-    mergeable: "MERGEABLE",
-    headRefName: branch,
-    headRefOid: runHeadSha,
-    mergedAt: null,
-  };
-
-  const supervisor = new Supervisor(config);
-  (supervisor as unknown as { github: Record<string, unknown> }).github = {
-    authStatus: async () => ({ ok: true, message: null }),
-    listAllIssues: async () => [issue],
-    listCandidateIssues: async () => [issue],
-    getIssue: async () => issue,
-    resolvePullRequestForBranch: async () => pr,
-    getChecks: async () => [],
-    getUnresolvedReviewThreads: async () => [],
-    getPullRequest: async () => pr,
-    getPullRequestIfExists: async () => null,
-    getMergedPullRequestsClosingIssue: async () => [],
-    closeIssue: async () => {
-      throw new Error("unexpected closeIssue call");
-    },
-    createPullRequest: async () => {
-      throw new Error("unexpected createPullRequest call");
-    },
-  };
-
-  const message = await supervisor.runOnce({ dryRun: false });
-  assert.match(message, /state=blocked/);
-
-  const persisted = JSON.parse(await fs.readFile(fixture.stateFile, "utf8")) as SupervisorStateFile;
-  const record = persisted.issues[String(issueNumber)];
-  assert.equal(record.state, "blocked");
-  assert.equal(record.blocked_reason, "manual_review");
-  assert.match(record.last_error ?? "", /requires manual review resolution before merge/);
-  assert.equal(record.last_failure_context?.category, "manual");
-  assert.match(record.last_failure_context?.summary ?? "", /requires manual review resolution before merge/);
-});
+import "./supervisor-pr-readiness.test";
+import "./supervisor-pr-review-blockers.test";
